% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gibsonsfuns-package.R
\docType{package}
\name{gibsonsfuns-package}
\alias{gibsonsfuns-package}
\alias{gibsonsfuns}
\title{The 'gibsonsfuns' package.}
\description{
A DESCRIPTION OF THE PACKAGE
}
\references{
Stan Development Team (NA). RStan: the R interface to Stan. R package version 2.32.7. https://mc-stan.org
}
\author{
\strong{Maintainer}: Yvan Richard \email{yvan@dragonfly.co.nz}

Authors:
\itemize{
  \item Yvan Richard \email{yvan@dragonfly.co.nz}
}

}
